Add RepositoryServiceProvider to Composer
The Commit removed an unnecessary GenerateLogicFactoryStub and added a RepositoryServiceProvider in its place in the laravel providers' list within Composer. The implementation of RepositoryServiceProvider enhances the registration process of service implementations. The introduction of the BaseLogic class simplifies the autowiring process and improves code readability via the handling of property bindings using ReflectionAttribute.